ROC-2161: moved interest* classes to features/claim/form/models
