+++ src/main/app/claims/claimModelConverter.ts
@@ -20,8 +20,8 @@ import InterestDate from 'app/claims/models/interestDate'
 import { Address } from 'claims/models/address'
 import { Address as AddressForm } from 'forms/models/address'
 import { ClaimAmountBreakdown } from 'features/claim/form/models/claimAmountBreakdown'
-import DraftInterestDate from 'app/forms/models/interestDate'
-import InterestDateType from 'app/common/interestDateType'
+import { InterestDate as DraftInterestDate } from 'claim/form/models/interestDate'
+import { InterestDateType } from 'app/common/interestDateType'
 import { StatementOfTruth } from 'claims/models/statementOfTruth'
 import StringUtils from 'utils/stringUtils'
