+++ src/main/app/claims/models/claim.ts
@@ -3,7 +3,7 @@ import { Serializable } from 'models/serializable'
 import { Moment } from 'moment'
 import ClaimData from 'app/claims/models/claimData'
 import { MomentFactory } from 'common/momentFactory'
-import InterestDateType from 'app/common/interestDateType'
+import { InterestDateType } from 'app/common/interestDateType'
 import { calculateInterest } from 'app/common/calculateInterest'
 import * as config from 'config'
 import * as toBoolean from 'to-boolean'
