+++ src/main/app/claims/models/claimData.ts
@@ -1,7 +1,7 @@
 import { Serializable } from 'app/models/serializable'
 import { ClaimAmountBreakdown } from 'features/claim/form/models/claimAmountBreakdown'
 import InterestDate from 'app/claims/models/interestDate'
-import Interest from 'app/forms/models/interest'
+import { Interest } from 'features/claim/form/models/interest'
 import { Party } from 'claims/models/details/yours/party'
 import { Individual as ClaimantAsIndividual } from 'claims/models/details/yours/individual'
 import { Company as ClaimantAsCompany } from 'claims/models/details/yours/company'
