+++ src/main/app/common/calculateInterest.ts
@@ -1,4 +1,4 @@
-import Interest, { InterestType } from 'forms/models/interest'
+import { Interest, InterestType } from 'claim/form/models/interest'
 import { Moment } from 'moment'
 import { MomentFactory } from 'common/momentFactory'
