+++ src/main/app/drafts/models/draftClaim.ts
@@ -1,8 +1,8 @@
 import { Serializable } from 'models/serializable'
 import Claimant from 'drafts/models/claimant'
-import { ClaimAmountBreakdown } from '../../../features/claim/form/models/claimAmountBreakdown'
-import Interest from 'forms/models/interest'
-import InterestDate from 'forms/models/interestDate'
+import { ClaimAmountBreakdown } from 'claim/form/models/claimAmountBreakdown'
+import { Interest } from 'claim/form/models/interest'
+import { InterestDate } from 'claim/form/models/interestDate'
 import Reason from 'forms/models/reason'
 import * as uuid from 'uuid'
 import { Defendant } from 'app/drafts/models/defendant'
