+++ src/main/app/pdf/mappers/interestMapper.ts
@@ -2,9 +2,9 @@ import { MomentFormatter } from 'utils/momentFormatter'
 import { NumberFormatter } from 'utils/numberFormatter'
 import ClaimData from 'claims/models/claimData'
 import { calculateInterest } from 'app/common/calculateInterest'
-import { InterestType } from 'app/forms/models/interest'
+import { InterestType } from 'features/claim/form/models/interest'
 import Claim from 'app/claims/models/claim'
-import InterestDateType from 'app/common/interestDateType'
+import { InterestDateType } from 'app/common/interestDateType'
 import { Moment } from 'moment'
 
 export class InterestMapper {
@@ -14,7 +14,8 @@ export class InterestMapper {
     const isNoInterest: boolean = claimData.interest.type === InterestType.NO_INTEREST
     if (isNoInterest) {
       return undefined
-    } {
+    }
+    {
       const interestDate = claimData.interestDate.type === InterestDateType.SUBMISSION ? claim.createdAt : claimData.interestDate.date
       return {
         rate: claimData.interest.rate,
