+++ src/main/features/ccj/routes/paid-amount-summary.ts
@@ -5,8 +5,8 @@ import { Paths } from 'ccj/paths'
 import { ErrorHandling } from 'common/errorHandling'
 import Claim from 'claims/models/claim'
 import { MomentFactory } from 'common/momentFactory'
-import { InterestType } from 'forms/models/interest'
-import InterestDateType from 'app/common/interestDateType'
+import { InterestType } from 'claim/form/models/interest'
+import { InterestDateType } from 'app/common/interestDateType'
 import { Moment } from 'moment'
 import { calculateInterest } from 'app/common/calculateInterest'
