+++ src/main/features/claim/routes/check-and-send.ts
@@ -5,9 +5,9 @@ import { FormValidator } from 'forms/validation/formValidator'
 import StatementOfTruth from 'forms/models/statementOfTruth'
 import FeesClient from 'fees/feesClient'
 import ClaimAmountTotal from 'forms/models/claimInterestTotal'
-import InterestDateType from 'app/common/interestDateType'
+import { InterestDateType } from 'app/common/interestDateType'
 import { claimAmountWithInterest, interestAmount } from 'utils/interestUtils'
-import { InterestType } from 'forms/models/interest'
+import { InterestType } from 'claim/form/models/interest'
 import { PartyType } from 'app/common/partyType'
 import AllClaimTasksCompletedGuard from 'claim/guards/allClaimTasksCompletedGuard'
 import { IndividualDetails } from 'forms/models/individualDetails'
