+++ src/main/features/claim/routes/interest-date.ts
@@ -1,12 +1,11 @@
 import * as express from 'express'
-
 import { Paths } from 'claim/paths'
-
 import { ErrorHandling } from 'common/errorHandling'
 import { Form } from 'forms/form'
 import { FormValidator } from 'forms/validation/formValidator'
-import InterestDate from 'forms/models/interestDate'
+import { InterestDate } from 'claim/form/models/interestDate'
 import { DraftService } from 'common/draft/draftService'
+
 function renderView (form: Form<InterestDate>, res: express.Response): void {
   res.render(Paths.interestDatePage.associatedView, { form: form })
 }
