+++ src/main/features/claim/routes/total.ts
@@ -1,6 +1,6 @@
 import * as express from 'express'
 import { Paths } from 'claim/paths'
-import { InterestType } from 'app/forms/models/interest'
+import { InterestType } from 'features/claim/form/models/interest'
 import InterestTotal from 'forms/models/claimInterestTotal'
 import { claimAmountWithInterest, interestAmount } from 'app/utils/interestUtils'
 import FeesClient from 'fees/feesClient'
