+++ src/main/features/testing-support/models/updateResponseDeadlineRequest.ts
@@ -2,7 +2,7 @@ import { LocalDate } from 'forms/models/localDate'
 import { IsValidYearFormat } from 'forms/validation/validators/isValidYearFormat'
 import { IsValidLocalDate } from 'forms/validation/validators/isValidLocalDate'
 import { IsDefined, ValidateNested } from 'class-validator'
-import { ValidationErrors as DateValidationErrors } from 'forms/models/interestDate'
+import { ValidationErrors as DateValidationErrors } from 'claim/form/models/interestDate'
 import { IsNotBlank } from 'forms/validation/validators/isBlank'
 
 class ValidationErrors {
