+++ src/test/app/drafts/models/interestDate.ts
@@ -1,10 +1,12 @@
 import { expect } from 'chai'
 
-import InterestDate from 'app/forms/models/interestDate'
-import InterestDateType from 'app/common/interestDateType'
+import { InterestDate } from 'claim/form/models/interestDate'
+import { InterestDateType } from 'app/common/interestDateType'
 
 describe('Interest', () => {
+
   describe('isCompleted', () => {
+
     it('should return true when interest type is submission', () => {
       const interest: InterestDate = new InterestDate(InterestDateType.SUBMISSION)
       expect(interest.isCompleted()).to.equal(true)
