+++ src/test/app/pdf/mappers/interestMapper.ts
@@ -1,7 +1,7 @@
 import { expect } from 'chai'
 
-import { InterestType } from 'app/forms/models/interest'
-import InterestDateType from 'app/common/interestDateType'
+import { InterestType } from 'claim/form/models/interest'
+import { InterestDateType } from 'app/common/interestDateType'
 import Claim from 'app/claims/models/claim'
 import { MomentFactory } from 'common/momentFactory'
 import { InterestMapper } from 'app/pdf/mappers/interestMapper'
