+++ src/test/data/draft/claimDraft.ts
@@ -1,6 +1,6 @@
 import { individualDetails } from './partyDetails'
-import { InterestType } from 'forms/models/interest'
-import InterestDateType from 'app/common/interestDateType'
+import { InterestType } from 'claim/form/models/interest'
+import { InterestDateType } from 'app/common/interestDateType'
 
 export const claimDraft = {
   externalId: 'fe6e9413-e804-48d5-bbfd-645917fc46e5',
