+++ src/test/http-mocks/draft-store.ts
@@ -6,7 +6,7 @@ import * as HttpStatus from 'http-status-codes'
 import { ResponseType } from 'response/form/models/responseType'
 import { FreeMediationOption } from 'response/form/models/freeMediation'
 import { MoreTimeNeededOption } from 'response/form/models/moreTimeNeeded'
-import { InterestType } from 'app/forms/models/interest'
+import { InterestType } from 'app/../../main/features/claim/form/models/interest'
 import { Defendant } from 'app/drafts/models/defendant'
 import Claimant from 'app/drafts/models/claimant'
 import DraftClaim from 'app/drafts/models/draftClaim'
@@ -18,8 +18,8 @@ import DateOfBirth from 'app/forms/models/dateOfBirth'
 import { LocalDate } from 'forms/models/localDate'
 import { ClaimAmountBreakdown } from 'claim/form/models/claimAmountBreakdown'
 import { ClaimAmountRow } from 'claim/form/models/claimAmountRow'
-import Interest from 'forms/models/interest'
-import InterestDate from 'forms/models/interestDate'
+import { Interest } from 'claim/form/models/interest'
+import { InterestDate } from 'claim/form/models/interestDate'
 import Reason from 'forms/models/reason'
 import { ResponseDraft } from 'response/draft/responseDraft'
 import Email from 'app/forms/models/email'
